+++ CHANGELOG.md
@@ -40,3 +40,4 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - User State
 - Keycloak profile
 - Bridgehead for profile with only one bridgehead in project view
+- Project configuration